+++ src/main/java/com/civilink/civilink_project_management/repositories/MainTaskRepository.java
@@ -4,7 +4,9 @@ import com.civilink.civilink_project_management.entities.MainTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
+
 @Repository
-public interface MainTaskRepository extends JpaRepository<MainTask, Integer> {
+public interface MainTaskRepository extends JpaRepository<MainTask, Long> {
 
 }
