+++ src/main/java/com/civilink/civilink_project_management/repositories/SubTaskRepository.java
@@ -4,6 +4,8 @@ import com.civilink.civilink_project_management.entities.SubTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
+
 @Repository
-public interface SubTaskRepository extends JpaRepository<SubTask, Integer> {
+public interface SubTaskRepository extends JpaRepository<SubTask, Long> {
 }
