+++ src/main/java/com/civilink/civilink_project_management/services/RetrieveMainTasksService.java
@@ -3,7 +3,8 @@ package com.civilink.civilink_project_management.services;
 import com.civilink.civilink_project_management.dtos.responses.ResponseMainTaskDto;
 import java.util.List;
 
+
 public interface RetrieveMainTasksService {
      List<ResponseMainTaskDto> getAllMainTasks();
-     ResponseMainTaskDto getMainTaskById(Integer id);
+     ResponseMainTaskDto getMainTaskById(Long id);
 }
