+++ src/main/java/com/civilink/civilink_project_management/services/RetrieveSubTasksService.java
@@ -2,8 +2,9 @@ package com.civilink.civilink_project_management.services;
 
 import com.civilink.civilink_project_management.dtos.responses.ResponseSubTaskDto;
 import java.util.List;
+import java.util.UUID;
 
 public interface RetrieveSubTasksService {
     List<ResponseSubTaskDto> getAllSubTasks();
-    ResponseSubTaskDto getSubTaskById(Integer id);
+    ResponseSubTaskDto getSubTaskById(Long id);
 }
